Add explicit height to the editor image only
The resizing tools we use size a parent div on the image. The height: auto responsive CSS we inherit from up higher messes with this. So this push adds an explicit 100% height on resized images.

Only missing piece is setting the is-resized property on the image block, when you manually input numbers.